refactor: reorganize FastAPI application into modular structure
- Extract Pydantic models to app/models package
- Move repository interfaces and implementations to app/repositories
- Separate Redis cache and MealDB services into app/services
- Organize API routes using FastAPI APIRouter in app/routers
- Implement dependency injection in app/dependencies module
- Refactor main.py to use clean import structure
- Maintain all existing functionality and API compatibility
- Update test imports to work with new module structure
- Add Redis dependency for cache service functionality

This refactor improves code organization, maintainability, and follows
Python/FastAPI best practices while preserving all existing features.
