+++ app/__init__.py
@@ -1,0 +1,6 @@
+# Recipe Discovery API Package
+"""
+A FastAPI application for recipe discovery with Redis caching and multiple data sources.
+"""
+
+__version__ = "1.0.0"
