+++ app/models/__init__.py
@@ -1,0 +1,4 @@
+# Models package
+from .recipe import Recipe, RecipeRequest, RecipeResponse
+
+__all__ = ["Recipe", "RecipeRequest", "RecipeResponse"]
