+++ app/routers/__init__.py
@@ -1,0 +1,5 @@
+# Routers package
+from .health import router as health_router
+from .recipes import router as recipes_router
+
+__all__ = ["health_router", "recipes_router"]
