+++ app/services/__init__.py
@@ -1,0 +1,5 @@
+# Services package
+from .cache import RedisCache
+from .mealdb import MealDBService
+
+__all__ = ["RedisCache", "MealDBService"]
